+++ CHANGELOG.md
@@ -6,6 +6,7 @@
 - handle drafts from mailto links in scanned QR #3492
 - do not overflow ratelimiter leaky bucket #3496
 - (AEAP) Add device message after you changed your address #3505
+- (AEAP) Revert #3491, instead only replace contacts in verified groups #3510
 
 ### Fixes
 - don't squash text parts of NDN into attachments #3497